Update to move server crates into backend.
Add core crate to share data structs between backend and frontend.
