--- Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-authors = ["nathanjent <nathanjent@gmail.com>"]
-name = "nathanjent"
-version = "0.1.0"
-
-[dependencies]
-dotenv = "0.8.0"
-http = "0.1.5"
-serde_derive = "1.0.11"
-
-[dependencies.backend]
-path = "backend"
-version = "0.1.0"
-
-[dependencies.common]
-path = "common"
-version = "0.1.0"
-
-[dependencies.diesel]
-features = ["mysql"]
-version = "0.16.0"
-
-[dependencies.diesel_codegen]
-features = ["mysql"]
-version = "0.16.0"
-
-[dependencies.frontend]
-path = "frontend"
-version = "0.1.0"
-
-[dependencies.http_router]
-path = "http_router"
-version = "0.1.0"
-
-[workspace]
-members = ["common", "frontend", "backend", "http_router", "query_parse"]
