+++ backend/common/src/static_content.rs
