+++ bootstrap.sh
@@ -76,5 +76,3 @@ echo "${DBSETUP}" > /vagrant/db_setup.sql
 mysql -uroot -p"${PASSWORD}" < /vagrant/db_setup.sql
 rm /vagrant/db_setup.sql
 
-# the username and password should match the DATABASE_URL variable in .env for diesel
-#echo DATABASE_URL=mysql://nathanjent:firetruck@localhost:3306/test > /vagrant/.env
