+++ core/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "core"
+version = "0.1.0"
+authors = ["nathanjent <nathanjent@gmail.com>"]
+
+[dependencies]
