+++ frontend/Cargo.toml
@@ -6,6 +6,7 @@ version = "0.1.0"
 [dependencies]
 query_params = "0.1.0"
 
-[dependencies.common]
-path = "../common"
+# Shared with backend
+[dependencies.core]
+path = "../core"
 version = "0.1.0"
